requirements.txt
openai
openpyxl
numpy
tiktoken
unstructured
uvicorn
jinja2
fastapi
python-dotenv
requests
apscheduler
pydantic
starlette
python-multipart
bs4
httpcore
pyyaml